+++ src/dodal/devices/areadetector/plugins/MJPG.py
