+++ src/dodal/devices/oav/snapshots/snapshot_with_grid.py
@@ -5,7 +5,7 @@ import aiofiles
 from ophyd_async.core import soft_signal_rw
 from PIL.Image import Image
 
-from dodal.devices.areadetector.plugins.MJPG_async import MJPG
+from dodal.devices.areadetector.plugins.MJPG import MJPG
 from dodal.devices.oav.snapshots.grid_overlay import (
     add_grid_border_overlay_to_image,
     add_grid_overlay_to_image,
